Move parameter renderer into the new render library directory
